Enabling UnsafeTypeCastWarningLevel for NetCore
#jira none
#preflight 6414df1842618c85c4d59f38
#rb ryan.gerleve

[CL 24697341 by Brian Bekich in ue5-main branch]

+++ Engine/Source/Runtime/Net/Core/Private/Net/Core/PushModel/PushModel.cpp
@@ -219,7 +219,7 @@ namespace UEPushModelPrivate
 			if (LIKELY(PerObjectStates.IsValidIndex(ObjectIndex)))
 			{
 				// The macros will take care of filtering out invalid objects, so we don't need to check here.
-				PerObjectStates[ObjectIndex].MarkPropertyDirty(RepIndex);
+				PerObjectStates[ObjectIndex].MarkPropertyDirty(static_cast<uint16>(RepIndex));
 			}
 		}
 
@@ -229,9 +229,9 @@ namespace UEPushModelPrivate
 			if (LIKELY(PerObjectStates.IsValidIndex(ObjectIndex)))
 			{
 				FPushModelPerObjectState& ObjectState = PerObjectStates[ObjectIndex];
-				for (int RepIndex = StartRepIndex; RepIndex <= EndRepIndex; ++RepIndex)
+				for (int32 RepIndex = StartRepIndex; RepIndex <= EndRepIndex; ++RepIndex)
 				{
-					ObjectState.MarkPropertyDirty(RepIndex);
+					ObjectState.MarkPropertyDirty(static_cast<uint16>(RepIndex));
 				}
 			}
 		}
